backend/requirements.txt
Flask
psycopg2-binary
redis
gunicorn